Re-organize conversation code: - Rename HumanMessage, AIMessage and SystemMessage to Prompt, Response, Context - Reorganize conversation classes: put everything conversatoin related directory
