--- lib/langchain/ai_message.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Langchain
-  class AIMessage < Message
-    def type
-      "ai"
-    end
-  end
-end
