+++ lib/langchain/conversation/context.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Langchain
+  class Conversation
+    class Context < Message
+      def type
+        "system"
+      end
+    end
+  end
+end
