+++ lib/langchain/conversation/prompt.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Langchain
+  class Conversation
+    class Prompt < Message
+      def type
+        "human"
+      end
+    end
+  end
+end
