--- lib/langchain/human_message.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Langchain
-  class HumanMessage < Message
-    def type
-      "human"
-    end
-  end
-end
