+++ lib/langchain/llm/cohere.rb
@@ -70,7 +70,7 @@ module Langchain::LLM
     # Cohere does not have a dedicated chat endpoint, so instead we call `complete()`
     def chat(...)
       response_text = complete(...)
-      Langchain::AIMessage.new(response_text)
+      ::Langchain::Conversation::Response.new(response_text)
     end
 
     # Generate a summary in English for a given text
